Update main.cpp
Added function for being able to convert srm saves to sav saves and vise-versa on the fly for the same game name but played in either Nintendo Switch Online Virtual Console vs independant emulator. Key motive will be to be able to play games in an emulator that supposed fast forward functionality normally, but then ale to play that same save file online with friends normally.